backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
pydantic>=2.9.0
python-multipart>=0.0.7
python-dotenv>=1.0.0
aiosqlite>=0.20.0
email-validator>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-decouple>=3.8
psycopg2-binary>=2.9.0
